Solution to day 3 part 2 + unit test. Change interface so that solvePartTwo returns long as this is needed for some puzzles

+++ src/main/java/net/lsmith946/adventofcode/problems2020/Day1.java
@@ -41,7 +41,7 @@ public class Day1 implements Puzzle {
     }
 
     @Override
-    public int solvePartTwo() {
+    public long solvePartTwo() {
         // part 2 is to find three integers from the list which sum to 2020
         // and then find the product of those three integers
         int product = 0;
